ggml-quants : faster sorting for cumulative search
Even when doing the full search with both signs, IQ4_NL with imatrix
is taking only 82% of the time it took in the original grid search,
instead of taking 198% of the time with the heap sort.
The sorting algorithm really was the bottleneck in the incremental search.

This now uses an hybrid of a counting sort with interpolated buckets,
and insertion sort.  So by mixing a non-comparative partial sort with
an adaptive comparative sort, this is much faster than any purely
comparative sort.